requirements.txt
Flask==3.1.0
flask-cors==5.0.1
pymongo==4.11.3
python-dotenv==1.1.0
PyJWT==2.10.1
bcrypt==4.3.0
requests==2.32.3
gunicorn==23.0.0
certifi
